adding heuristic to get signess of operations from CFA.
This heuritic looks for the signess of the original type of a symbol and
   sets the flag according to the found signess.
We do not need the signess of all symbols, so we do not match back from
any String, but only from the most common String-representation of a
variable, i.e. a name with an SSA-index like "name@index".

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@16974 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c