BVConverter tries to get the signess of a BV-operation.
Adding heuristic to get the correct signess for a BV-operations. Before
we assumed only "signed operators". Now we try to get the signess for
all operands and merge them (similar to CBinaryExpressionBuilder, but
much simpler).
We can not guarantee to get the correct signess, because predicates come
from interpolants and are not self-made formulas. Thus any signess is
possible in an predicate. However some tests show better results with
this change.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@16975 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c