SymEx refinement: strongest-post op mirrors VA-ConstraintsCPA relation correctly.
ConstraintsCPA creates new state in strengthening with VA state.
VA is then strengthened with constraints state, but the one before strengthening (which is the state of one iteration before, as constraints transfer doesn't do anything).
This means that the 'real' strengthening of VA with the constraints state happens one iteration later.

Represent this by first strengthening the value analysis state and then performing other computations, instead of strengthening at the end of the strongest-post operation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@17127 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c